Toliau mažinti props: leaf komponentai patys pasiima duomenis per context
RouteStats, RouteStepsList, NavigationInstructionBanner, NavigationProgressBar,
NavigationStepsSheet dabar patys skaito duomenis (instructions, routeLine,
selectedRouteProfile, remainingDistance/Time ir pan.) iš RouteProvider
context'ų, o ne gauna juos per props iš tėvinio komponento. Props liko tik
tikri elgsenos callback'ai, kurie skiriasi priklausomai nuo iškvietimo vietos
(pvz. NavigationStepsSheet po paspaudimo dar uždaro sąrašą, o RouteDetails -
ne), arba lokali UI būsena (onOpenList).

MobileNavigationView dėl to sutrumpėjo 85 → 34 eilučių.

Taip pat išskaidytas RouteLayer.tsx (203 eil.) - jame buvo 3 skirtingos
kameros valdymo useEffect logikos + 4 žymeklių tipai + 2 statuso pranešimai
viename faile:
- useNavigationCamera hook'as - kameros sekimas/pasukimas navigacijos metu,
  bearing reset pabaigus, fitBounds naujam maršrutui.
- RouteMarkers - visi žemėlapio žymekliai (nav taškas/start/end/highlight).
- RouteStatusToast - "Skaičiuojamas maršrutas.../Maršruto klaida" pranešimai.

RouteLayer.tsx: 203 → 47 eilučių.

Jokių funkcinių pakeitimų - patikrinta Playwright regresija (progresas,
pauzė/tęsimas, atvykimas, nuklydimo perskaičiavimas, mobili navigacija su
kameros pasukimu/priartinimu).
