+++ src/components/route/MobileNavigationView.tsx
@@ -1,14 +1,8 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { useActiveInstruction } from "@/hooks/use-active-instruction";
-import { useRouteMapFocus } from "@/hooks/use-route-map-focus";
 import { useIsMobile } from "@/hooks/useIsMobile";
-import {
-  useNavigationProgress,
-  useRoute,
-  useRouteResult,
-} from "@/providers/RouteProvider";
+import { useRoute } from "@/providers/RouteProvider";
 import { NavigationInstructionBanner } from "./NavigationInstructionBanner";
 import { NavigationProgressBar } from "./NavigationProgressBar";
 import { NavigationStepsSheet } from "./NavigationStepsSheet";
@@ -17,24 +11,7 @@ import { NavigationStepsSheet } from "./NavigationStepsSheet";
 export function MobileNavigationView() {
   const isMobile = useIsMobile();
   const [showFullList, setShowFullList] = useState(false);
-  const {
-    routingMode,
-    navigationMode,
-    selectedRouteProfile,
-    routeStart,
-    routeEnd,
-    setNavigationMode,
-  } = useRoute();
-  const { routeLine, instructions } = useRouteResult();
-  const progress = useNavigationProgress();
-  const { flyToInstruction, flyToEndpoint } = useRouteMapFocus(routeLine);
-  const {
-    currentIndex,
-    activeInstruction,
-    nextInstruction,
-    liveDistanceToNext,
-    remainingTime,
-  } = useActiveInstruction(instructions, routeLine);
+  const { routingMode, navigationMode } = useRoute();
 
   const isActive = navigationMode && isMobile && routingMode;
 
@@ -45,41 +22,13 @@ export function MobileNavigationView() {
   if (!isActive) return null;
 
   if (showFullList) {
-    return (
-      <NavigationStepsSheet
-        instructions={instructions}
-        routeStartName={routeStart?.properties?.name}
-        routeEndName={routeEnd?.properties?.name}
-        selectedRouteProfile={selectedRouteProfile}
-        currentIndex={currentIndex}
-        liveDistanceToNext={liveDistanceToNext}
-        onInstructionClick={(step) => {
-          flyToInstruction(step);
-          setShowFullList(false);
-        }}
-        onStartEndClick={flyToEndpoint}
-        onClose={() => setShowFullList(false)}
-      />
-    );
+    return <NavigationStepsSheet onClose={() => setShowFullList(false)} />;
   }
 
   return (
     <>
-      <NavigationInstructionBanner
-        activeInstruction={activeInstruction}
-        activeDistance={
-          liveDistanceToNext ?? activeInstruction?.distance ?? null
-        }
-        nextInstruction={nextInstruction}
-        arrived={progress.arrived}
-        onOpenList={() => setShowFullList(true)}
-      />
-      <NavigationProgressBar
-        remainingDistance={progress.remainingDistance}
-        remainingTime={remainingTime}
-        locationError={!!progress.error}
-        onStop={() => setNavigationMode(false)}
-      />
+      <NavigationInstructionBanner onOpenList={() => setShowFullList(true)} />
+      <NavigationProgressBar />
     </>
   );
 }
